userservice/requirement.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
PyJWT
django-cors-headers
django-mysql
mysqlclient
pycryptodome